// src/app/shared/models/acquirer.model.ts
export interface AcquirerModel {
  acquiringInstitutionCode: string;
  city: string;
  code: string;
  country: string;
  hsmKeyName: string;
  hsmTransitKeyName: string;
  name: string;
  region: string;
  street: string;
  zip: string;
}

export interface AcquirerPredicateObject {
  acquiringInstitutionCode: string;
  city: string;
  code: string;
  country: string;
  hsmKeyName: string;
  hsmTransitKeyName: string;
  name: string;
  region: string;
  street: string;
  zip: string;
}